backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
beautifulsoup4==4.12.2
requests==2.31.0
aiohttp==3.9.1
lxml==4.9.3
python-dotenv==1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
redis==5.0.1
python-multipart==0.0.6
httpx==0.25.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
transformers==4.35.2
torch>=2.1.0
pillow>=10.0.0
anthropic>=0.39.0
tensorflow>=2.13.0
xhtml2pdf>=0.2.11
jinja2>=3.1.0
